[BAH-835]extract out the SaleShop from sale_order.py and created new file shop.py, and StockMove from stock_picking.py and set in stock_move.py file.
