+++ bahmni_sale/models/__init__.py
@@ -6,3 +6,4 @@ import sale_config_settings
 import res_company
 import pos
 import account_invoice
+import shop
