+++ bahmni_stock/models/stock_move.py
@@ -26,3 +26,10 @@ class StockMove(models.Model):
     
     stock_picking_time = fields.Datetime(compute=_get_picking_time,
                                          string="Stock Picking_time", store=True)
+    @api.model
+    def create(self,vals):
+        if vals.get('origin'):
+            sale_order = self.env['sale.order'].search([('name','=',str(vals.get('origin')))])
+            if any(sale_order) and sale_order.location_id:
+                vals.update({'location_id':sale_order.location_id.id})
+        return super(StockMove,self).create(vals)
